flask-server/requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.1
flask-ldap-login==0.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
ldap3==2.9.1
MarkupSafe==2.1.5
pyasn1==0.6.0
pycryptodome==3.20.0
SQLAlchemy==2.0.30
typing_extensions==4.12.1
Werkzeug==3.0.3
wfastcgi==3.0.0
